+++ ChangeLog.md
@@ -5,6 +5,7 @@
 **Released: WiP**
 
 - Made small improvements to article list refreshing.
+  ([#21](https://github.com/davep/oldnews/pull/21))
 
 ## v0.2.0