+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # OldNews ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Made small improvements to article list refreshing.
+  ([#21](https://github.com/davep/oldnews/pull/21))
+
 ## v0.2.0
 
 **Released: 2026-01-17**