+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Made small improvements to article list refreshing.
   ([#21](https://github.com/davep/oldnews/pull/21))
 - Moved the sync code out into its own class.
+  ([#22](https://github.com/davep/oldnews/pull/22))
 
 ## v0.2.0