+++ ChangeLog.md
@@ -9,6 +9,7 @@
 - Moved the sync code out into its own class.
   ([#22](https://github.com/davep/oldnews/pull/22))
 - Made some cosmetic tweaks to the article view.
+  ([#23](https://github.com/davep/oldnews/pull/23))
 
 ## v0.2.0