+++ ChangeLog.md
@@ -14,14 +14,17 @@
   data (and optionally force a logout).
   ([#27](https://github.com/davep/oldnews/pull/27))
 - Removed the `--licence` command line switch.
+  ([#29](https://github.com/davep/oldnews/pull/29))
 - Removed the `--bindings` command line switch.
+  ([#29](https://github.com/davep/oldnews/pull/29))
 - Removed the "?" parameter for the `--theme` command line switch.
+  ([#29](https://github.com/davep/oldnews/pull/29))
 - Added a `licence`/`license` command line command that is used to show the
-  application's licence.
+  application's licence. ([#29](https://github.com/davep/oldnews/pull/29))
 - Added `bindings` command line command that is used to list all bindable
-  commands.
+  commands. ([#29](https://github.com/davep/oldnews/pull/29))
 - Added a `themes` command that is used to list all themes available to the
-  application.
+  application. ([#29](https://github.com/davep/oldnews/pull/29))
 
 ## v0.2.0